[FLINK-23192][connector-hbase] Move HBase table connector options class
